+++ lib/request/sfWebRequest.class.php
@@ -70,6 +70,7 @@ class sfWebRequest extends sfRequest
       'http_port'       => null,
       'https_port'      => null,
       'default_format'  => null, // to maintain bc
+      'trust_proxy'     => true, // to maintain bc
     ), $options);
     parent::initialize($dispatcher, $parameters, $attributes, $options);
 
@@ -88,15 +89,15 @@ class sfWebRequest extends sfRequest
           break;
 
         case 'POST':
-          if (isset($_POST['sf_method']))
+          if (isset($postParameters['sf_method']))
           {
-            $this->setMethod(strtoupper($_POST['sf_method']));
+            $this->setMethod(strtoupper($postParameters['sf_method']));
             unset($postParameters['sf_method']);
           }
-          elseif (isset($_GET['sf_method']))
+          elseif (isset($this->getParameters['sf_method']))
           {
-            $this->setMethod(strtoupper($_GET['sf_method']));
-            unset($_GET['sf_method']);
+            $this->setMethod(strtoupper($this->getParameters['sf_method']));
+            unset($this->getParameters['sf_method']);
           }
           else
           {
@@ -138,9 +139,9 @@ class sfWebRequest extends sfRequest
     $this->postParameters = get_magic_quotes_gpc() ? sfToolkit::stripslashesDeep($postParameters) : $postParameters;
     $this->parameterHolder->add($this->postParameters);
 
-    if (isset($this->options['formats']))
+    if ($formats = $this->getOption('formats'))
     {
-      foreach ($this->options['formats'] as $format => $mimeTypes)
+      foreach ($formats as $format => $mimeTypes)
       {
         $this->setFormat($format, $mimeTypes);
       }
@@ -182,7 +183,7 @@ class sfWebRequest extends sfRequest
     $pathArray = $this->getPathInfoArray();
 
     // for IIS with rewrite module (IIFR, ISAPI Rewrite, ...)
-    if ('HTTP_X_REWRITE_URL' == $this->options['path_info_key'])
+    if ('HTTP_X_REWRITE_URL' == $this->getOption('path_info_key'))
     {
       $uri = isset($pathArray['HTTP_X_REWRITE_URL']) ? $pathArray['HTTP_X_REWRITE_URL'] : '';
     }
@@ -225,9 +226,9 @@ class sfWebRequest extends sfRequest
     {
       list($host, $port) = explode(':', $host, 2);
     }
-    else if (isset($this->options[$protocol.'_port']))
+    else if ($protocolPort = $this->getOption($protocol.'_port'))
     {
-      $port = $this->options[$protocol.'_port'];
+      $port = $protocolPort;
     }
     else if (isset($pathArray['SERVER_PORT']))
     {
@@ -238,7 +239,7 @@ class sfWebRequest extends sfRequest
     // a secure one and whether the introspected port matches the standard one
     if ($this->isForwardedSecure())
     {
-      $port = isset($this->options['https_port']) && self::PORT_HTTPS != $this->options['https_port'] ? $this->options['https_port'] : null;
+      $port = self::PORT_HTTPS != $this->getOption('https_port') ? $this->getOption('https_port') : null;
     }
     elseif (($secure && self::PORT_HTTPS == $port) || (!$secure && self::PORT_HTTP == $port))
     {
@@ -260,7 +261,7 @@ class sfWebRequest extends sfRequest
     $pathArray = $this->getPathInfoArray();
 
     // simulate PATH_INFO if needed
-    $sf_path_info_key = $this->options['path_info_key'];
+    $sf_path_info_key = $this->getOption('path_info_key');
     if (!isset($pathArray[$sf_path_info_key]) || !$pathArray[$sf_path_info_key])
     {
       if (isset($pathArray['REQUEST_URI']))
@@ -298,11 +299,16 @@ class sfWebRequest extends sfRequest
     return $pathInfo;
   }
 
+  /**
+   * Returns the relative url root if defined computed with script name if defined
+   *
+   * @return string The path info prefix
+   */
   public function getPathInfoPrefix()
   {
     $prefix = $this->getRelativeUrlRoot();
 
-    if (!isset($this->options['no_script_name']) || !$this->options['no_script_name'])
+    if (!$this->getOption('no_script_name'))
     {
       $scriptName = $this->getScriptName();
       $prefix = null === $prefix ? $scriptName : $prefix.'/'.basename($scriptName);
@@ -311,22 +317,42 @@ class sfWebRequest extends sfRequest
     return $prefix;
   }
 
+  /**
+   * Gets GET parameters from request
+   *
+   * @return array
+   */
   public function getGetParameters()
   {
     return $this->getParameters;
   }
 
+  /**
+   * Gets POST parameters from request
+   *
+   * @return array
+   */
   public function getPostParameters()
   {
     return $this->postParameters;
   }
 
+  /**
+   * Gets REQUEST parameters from request
+   *
+   * @return array
+   */
   public function getRequestParameters()
   {
     return $this->requestParameters;
   }
 
-  public function addRequestParameters($parameters)
+  /**
+   * Add fixed REQUEST parameters
+   *
+   * @param array $parameters
+   */
+  public function addRequestParameters(array $parameters)
   {
     $this->requestParameters = array_merge($this->requestParameters, $parameters);
     $this->getParameterHolder()->add($parameters);
@@ -355,16 +381,14 @@ class sfWebRequest extends sfRequest
   {
     $pathArray = $this->getPathInfoArray();
 
-    if (isset($pathArray['HTTP_X_FORWARDED_HOST']))
+    if ($this->getOption('trust_proxy') && isset($pathArray['HTTP_X_FORWARDED_HOST']))
     {
       $elements = explode(',', $pathArray['HTTP_X_FORWARDED_HOST']);
 
       return trim($elements[count($elements) - 1]);
     }
-    else
-    {
-      return isset($pathArray['HTTP_HOST']) ? $pathArray['HTTP_HOST'] : '';
-    }
+
+    return isset($pathArray['HTTP_HOST']) ? $pathArray['HTTP_HOST'] : '';
   }
 
   /**
@@ -529,6 +553,13 @@ class sfWebRequest extends sfRequest
     return ($this->getHttpHeader('X_REQUESTED_WITH') == 'XMLHttpRequest');
   }
 
+  /**
+   * Gets the value of HTTP header
+   *
+   * @param string $nameThe HTTP header name
+   * @param string $prefix The HTTP header prefix
+   * @return string The value of HTTP header
+   */
   public function getHttpHeader($name, $prefix = 'http')
   {
     if ($prefix)
@@ -544,12 +575,12 @@ class sfWebRequest extends sfRequest
   }
 
   /**
-   * Gets a cookie value.
+   * Gets the value of a cookie.
    *
    * @param  string $name          Cookie name
    * @param  string $defaultValue  Default value returned when no cookie with given name is found
    *
-   * @return mixed
+   * @return string The cookie value
    */
   public function getCookie($name, $defaultValue = null)
   {
@@ -573,11 +604,11 @@ class sfWebRequest extends sfRequest
     $pathArray = $this->getPathInfoArray();
 
     return
-      (isset($pathArray['HTTPS']) && ('on' == strtolower($pathArray['HTTPS']) || 1 == $pathArray['HTTPS']))
+      (isset($pathArray['HTTPS']) && (('on' == strtolower($pathArray['HTTPS']) || 1 == $pathArray['HTTPS'])))
       ||
-      (isset($pathArray['HTTP_SSL_HTTPS']) && ('on' == strtolower($pathArray['HTTP_SSL_HTTPS']) || 1 == $pathArray['HTTP_SSL_HTTPS']))
+      ($this->getOption('trust_proxy') && isset($pathArray['HTTP_SSL_HTTPS']) && (('on' == strtolower($pathArray['HTTP_SSL_HTTPS']) || 1 == $pathArray['HTTP_SSL_HTTPS'])))
       ||
-      $this->isForwardedSecure()
+      ($this->getOption('trust_proxy') && $this->isForwardedSecure())
     ;
   }
 
@@ -602,13 +633,9 @@ class sfWebRequest extends sfRequest
   {
     if (null === $this->relativeUrlRoot)
     {
-      if (!isset($this->options['relative_url_root']))
+      if (!($this->relativeUrlRoot = $this->getOption('relative_url_root')))
       {
         $this->relativeUrlRoot = preg_replace('#/[^/]+\.php5?$#', '', $this->getScriptName());
-      }
-      else
-      {
-        $this->relativeUrlRoot = $this->options['relative_url_root'];
       }
     }
 
@@ -669,7 +696,7 @@ class sfWebRequest extends sfRequest
     if (!$this->pathInfoArray)
     {
       // parse PATH_INFO
-      switch ($this->options['path_info_array'])
+      switch ($this->getOption('path_info_array'))
       {
         case 'SERVER':
           $this->pathInfoArray =& $_SERVER;
@@ -752,7 +779,7 @@ class sfWebRequest extends sfRequest
   {
     if (null === $this->format)
     {
-      $this->setRequestFormat($this->getParameter('sf_format', $this->options['default_format']));
+      $this->setRequestFormat($this->getParameter('sf_format', $this->getOption('default_format')));
     }
 
     return $this->format;
@@ -794,7 +821,14 @@ class sfWebRequest extends sfRequest
     return $files;
   }
 
-  static protected function fixPhpFilesArray($data)
+  /**
+   * Fixes PHP files array
+   *
+   * @param array $data The PHP files
+   *
+   * @return array The fixed PHP files array
+   */
+  static protected function fixPhpFilesArray(array $data)
   {
     $fileKeys = array('error', 'name', 'size', 'tmp_name', 'type');
     $keys = array_keys($data);
@@ -922,40 +956,35 @@ class sfWebRequest extends sfRequest
   /**
    * Returns the client IP address that made the request.
    *
-   * @return string The client IP address
+   * @param  boolean $proxy Whether the current request has been made behind a proxy or not
+   *
+   * @return string Client IP(s)
    */
-  public function getClientIp()
+  public function getClientIp($proxy = true)
   {
-    $pathInfo = $this->getPathInfoArray();
-
-    if (isset($pathInfo["HTTP_CLIENT_IP"]))
+    if ($proxy)
     {
-      return $pathInfo["HTTP_CLIENT_IP"];
+      $pathInfo = $this->getPathInfoArray();
+
+      if (isset($pathInfo["HTTP_CLIENT_IP"]) && ($ip = $pathInfo["HTTP_CLIENT_IP"]))
+      {
+        return $ip;
+      }
+
+      if ($this->getOption('trust_proxy') && ($ip = $this->getForwardedFor()))
+      {
+        return isset($ip[0]) ? trim($ip[0]) : '';;
+      }
     }
 
-    return null;
+    return $this->getRemoteAddress();
   }
 
   /**
-   * Return the real IP (from proxies that passed the request or client if or remote addr)
+   * Check CSRF protection
    *
-   * @return array An array of real IP(s)
+   * @throws <b>sfValidatorErrorSchema</b> If an error occurs while validating the CRF protection for this sfRequest
    */
-  public function getRealIp()
-  {
-    if ($ip = $this->getForwardedFor())
-    {
-      return $ip;
-    }
-
-    if ($ip = $this->getClientIp())
-    {
-      return array($ip);
-    }
-
-    return array($this->getRemoteAddress());
-  }
-
   public function checkCSRFProtection()
   {
     $form = new BaseForm();
@@ -997,9 +1026,11 @@ class sfWebRequest extends sfRequest
     );
   }
 
+  /**
+   * Move symfony parameters to attributes (parameters prefixed with _sf_)
+   */
   protected function fixParameters()
   {
-    // move symfony parameters to attributes (parameters prefixed with _sf_)
     foreach ($this->parameterHolder->getAll() as $key => $value)
     {
       if (0 === stripos($key, '_sf_'))